feat!: Use shader instancing by using a Texture as a look up talbe
the ShaderStorageBuffer is not supported on WASM, of course it isn't...

This solution is actually quite nice, the only problem is that we need
to upload the lookup texture every frame, but that isn't too much of a
problem as we could put EVERYTHING we need to sync between CPU-GPU on
there (more than just the grass stuff).

So the overhead of sending it would be O(1), and the size of the image
could scale quite a bit until it becomes a problem.

So for a steam release of idle garden that should not be an issue.

Also, this runs butter smooth on WASM now as well, even when zooming all
the way out. Very nice.