fix!: Use old player sprite, rework grass to work with 32px tile size, various small things
Removed stuff like player blocking and shit. That would be necessary
when you have trees and shit, but we ain't gonna have that in the
Bachelor. I will have to figure something out for the steam build, but
that is when the time comes.

Actually, you are probably gonna plant seeds always right? Then maybe it
will be completely different anyways.